// frontend/src/app/lib/enums.ts
export enum DataSetName {
    ALL_RESOURCES = 'allResources',
    ONE_RESOURCE = 'oneResource',
    RESOURCE_HISTORY = 'resourceHistory'
}

export enum ProcessName {
    LOADING_ALL = 'loadingAll',
    LOADING_ONE = 'loadingOne',
    DELETING_ONE = 'deletingOne',
    UPDATING_ONE = 'updatingOne',
    ADDING_ONE = 'addingOne',
    LOADING_HISTORY = 'loadingHistory',
}

export enum ModalName {
    ADD_MODAL = 'addModal',
    UPDATE_MODAL = 'updateModal',
    DELETE_MODAL = 'deleteModal',
}

export enum Roles {
    FARMER = 'FARMER',
    REGULATION_CHECKER = 'REGULATION CHECKER',
    MANAGER = 'MANAGER',
    ADMINISTRATOR = 'ADMINISTRATOR',
}